refactor: address Gemini code review feedback
Apply improvements suggested by Gemini AI code review:

1. Use e.currentTarget instead of e.target in tab event handlers
   - More robust against future additions of child elements in buttons
   - Always references the button element where listener is attached

2. Add explanatory comments for evaluate() vs check() in tests
   - Clarifies why evaluate(el => el.click()) is necessary
   - Playwright's check() fails even with force:true for display:none elements

3. Extract shared test helper to eliminate code duplication
   - Create tests/e2e/test-helpers.js with switchToEffectTab function
   - Remove duplicate code from both app-webgl.test.js and app-webgpu.test.js
   - Improves maintainability - single source of truth for effect-to-tab mapping

All 14 tests passing.
