+++ apps/cybervision/tests/e2e/app-webgl.test.js
@@ -1,33 +1,5 @@
 import { test, expect } from '@playwright/test';
-
-// Helper function to switch to the tab containing a specific effect
-async function switchToEffectTab(page, effectValue) {
-  const effectToTab = {
-    'halftone': 'artistic',
-    'clustering': 'artistic',
-    'mosaic': 'artistic',
-    'kaleidoscope': 'artistic',
-    'pixelsort': 'artistic',
-    'original': 'distortion',
-    'edges': 'distortion',
-    'chromatic': 'distortion',
-    'glitch': 'distortion',
-    'thermal': 'distortion'
-  };
-
-  const tabName = effectToTab[effectValue];
-  if (tabName) {
-    const tabButton = page.locator(`button[data-tab="${tabName}"]`);
-    const isActive = await tabButton.evaluate(el => el.classList.contains('active'));
-
-    // Only click if tab is not already active
-    if (!isActive) {
-      await tabButton.click();
-      // Wait for tab content to become visible
-      await page.waitForTimeout(100);
-    }
-  }
-}
+import { switchToEffectTab } from './test-helpers.js';
 
 test.describe('CyberVision E2E - WebGL Path', () => {
   test('should load page with WebGL fallback', async ({ page }) => {
@@ -92,7 +64,9 @@ test.describe('CyberVision E2E - WebGL Path', () => {
     for (const effect of effects) {
       await switchToEffectTab(page, effect);
       const radio = page.locator(`input[value="${effect}"]`);
-      // Force the check using evaluate since radio is in a hidden tab
+      // Use evaluate() instead of check({ force: true }) because Playwright's
+      // check() still attempts to scroll into view even with force: true,
+      // which fails for elements in inactive tabs with display: none
       await radio.evaluate(el => el.click());
       await expect(radio).toBeChecked();
     }
@@ -148,7 +122,9 @@ test.describe('CyberVision E2E - WebGL Path', () => {
 
     for (const effect of effects) {
       await switchToEffectTab(page, effect);
-      // Force the check using evaluate since radio is in a hidden tab
+      // Use evaluate() instead of check({ force: true }) because Playwright's
+      // check() still attempts to scroll into view even with force: true,
+      // which fails for elements in inactive tabs with display: none
       await page.locator(`input[value="${effect}"]`).evaluate(el => el.click());
       // No need to wait - effect switching is synchronous in WebGL
     }
